// talkhub_frontend/src/app/api/model/updateUserDto.ts
/**
 * Talkhub API
 * This is a forum API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateUserDto { 
    /**
     * The user email
     */
    email: string;
    /**
     * The user name
     */
    name: string;
    /**
     * The user\'s role
     */
    role: UpdateUserDto.RoleEnum;
    /**
     * The user\'s bio
     */
    bio: string;
}
export namespace UpdateUserDto {
    export type RoleEnum = 'admin' | 'regular' | 'moderator';
    export const RoleEnum = {
        Admin: 'admin' as RoleEnum,
        Regular: 'regular' as RoleEnum,
        Moderator: 'moderator' as RoleEnum
    };
}